ok all converted to cpp standard, maybe some loss in precision at some point but we will take care of this later on

+++ generator.hpp
@@ -1,7 +1,7 @@
 #ifndef GENERATOR_H_
 #define GENERATOR_H_
 
-#include "layers.h"
+#include "layers.hpp"
 
 /* Minecraft versions */
 enum
@@ -199,7 +199,7 @@ void freeGenerator(LayerStack g);
 /* Calculates the minimum size of the buffers required to generate an area of
  * dimensions 'sizeX' by 'sizeZ' at the specified layer.
  */
-int calcRequiredBuf(Layer *layer, int areaX, int areaZ);
+int calcRequiredBuf(Layer *layer, long long areaX, long long areaZ);
 
 /* Allocates an amount of memory required to generate an area of dimensions
  * 'sizeX' by 'sizeZ' for the magnification of the current top layer.
@@ -208,8 +208,8 @@ int *allocCache(Layer *layer, int sizeX, int sizeZ);
 
 
 /* Set up custom layers. */
-void setupLayer(int scale, Layer *l, Layer *p, int s, void (*getMap)(Layer *layer, int *out, int x, int z, int w, int h));
-void setupMultiLayer(int scale, Layer *l, Layer *p1, Layer *p2, int s, void (*getMap)(Layer *layer, int *out, int x, int z, int w, int h));
+void setupLayer(int scale, Layer *l, Layer *p, int s, void (*getMap)(Layer *layer, int *out, long long x, long long z, int w, int h));
+void setupMultiLayer(int scale, Layer *l, Layer *p1, Layer *p2, int s, void (*getMap)(Layer *layer, int *out, long long x, long long z, int w, int h));
 
 
 /* Sets the world seed for the generator */
@@ -221,7 +221,7 @@ void applySeed(LayerStack *g, int64_t seed);
  * It is recommended that 'out' is allocated using allocCache() for the correct
  * buffer size.
  */
-void genArea(Layer *layer, int *out, int areaX, int areaZ, int areaWidth, int areaHeight);
+void genArea(Layer *layer, int *out, long long areaX, long long areaZ, int areaWidth, int areaHeight);
 
 
 #endif /* GENERATOR_H_ */
